+++ lib/src/center_on_location_update.dart
@@ -1,0 +1,5 @@
+enum CenterOnLocationUpdate {
+  never,
+  once,
+  always,
+}